refactored codes

+++ bloglist-backend/controllers/blogs.js
@@ -39,28 +39,29 @@ blogsRouter.get("/:id", async (request, response, next) => {
 
 blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
   const body = request.body;
+  console.log("body", body);
   // get user from request object
   const user = await request.user;
-  logger.info("user from request.user", user);
   // const token = getTokenFrom(request);
   // const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (body.title === undefined || body.url === undefined) {
     return response.status(400).json({ error: "title or url missing" });
   }
 
-  logger.info("user._id", user._id);
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes === undefined ? 0 : body.likes,
+    ...body,
     user: user._id,
   });
 
   const savedBlog = await blog.save();
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
-  response.status(201).json(savedBlog);
+
+  const blogToReturn = await Blog.findById(savedBlog._id).populate("user", {
+    username: 1,
+    name: 1,
+  });
+  response.status(201).json(blogToReturn);
 });
 
 blogsRouter.delete(
@@ -80,6 +81,14 @@ blogsRouter.delete(
     // }
 
     const userid = user._id;
+    if (!blog) {
+      return response.status(204).end();
+    }
+    if (blog.user && blog.user.toString() !== request.user.id) {
+      return response.status(401).json({
+        error: "only the creator can delete a blog",
+      });
+    }
     if (blog.user.toString() === userid.toString()) {
       await Blog.findByIdAndRemove(request.params.id);
       response.status(204).end();
@@ -91,35 +100,24 @@ blogsRouter.delete(
   }
 );
 
-blogsRouter.put(
-  "/:id",
-  middleware.userExtractor,
-  async (request, response, next) => {
-    const body = request.body;
-    const blog = await Blog.findById(request.params.id);
-    const user = request.user;
-    const userid = user._id;
+blogsRouter.put("/:id", async (request, response) => {
+  const body = request.body;
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user;
 
-    const newBlog = {
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes,
-    };
+  const newBlog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+  };
 
-    if (blog.user.toString() === userid.toString()) {
-      const updatedBlog = await Blog.findByIdAndUpdate(
-        request.params.id,
-        newBlog,
-        {
-          new: true,
-          runValidators: true,
-        }
-      );
-      response.status(200).json(updatedBlog);
-    }
-  }
-);
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, newBlog, {
+    new: true,
+    runValidators: true,
+  }).populate("user", { username: 1, name: 1 });
+  response.status(200).json(updatedBlog);
+});
 
 // comments
 blogsRouter.post("/:id/comments", async (request, response) => {
@@ -133,7 +131,7 @@ blogsRouter.post("/:id/comments", async (request, response) => {
   });
 
   const savedComment = await comment.save();
-  blog.comments = blog.comments.concat(savedComment._id);
+  blog.comments = blog.comments.concat(savedComment);
   await blog.save();
   response.status(201).json(savedComment);
 });
